+++ __tests__/e2e.ts
@@ -2,7 +2,7 @@ import * as exported from '../src/index';
 
 describe('e2e', () => {
   it('exports the correct functions', () => {
-    expect(exported.coerceError).toBeDefined();
+    expect(exported.maybeError).toBeDefined();
     expect(exported.trimTrace).toBeDefined();
   });
 });